refactor(environment): flatten structure
