+++ environment/logging.ts
@@ -1,6 +1,6 @@
 import { Workspace } from '@pulumi/azure-native/operationalinsights'
 
-import identity from '../identity'
+import identity from './identity'
 
 type Options = {
   identity: Awaited<ReturnType<typeof identity>>,
